Rename 'styles' folder to 'constants'
