+++ src/navigation/MainNavigator.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createBottomTabNavigator } from 'react-navigation';
 import { FontAwesome } from '@expo/vector-icons';
-import { colors } from 'styles/colors';
+import { colors } from 'constants/Colors';
 import TaskList from 'screens/TaskListScreen';
 import Profile from 'screens/ProfileScreen';
